Dev 1st sprint

+++ app/Http/Controllers/API/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Models\Department;
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Models\Product;
@@ -33,7 +34,9 @@ class ProductController extends BaseController
 
         $validator = Validator::make($input, [
             'name' => 'required',
-            'detail' => 'required'
+            'detail' => 'required',
+            'department_id' => 'required',
+            'category_id' => 'required',
         ]);
 
         if($validator->fails()){
@@ -66,12 +69,14 @@ class ProductController extends BaseController
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(Request $request, int $id)
     {
         $input = $request->all();
+
+        $product = Product::find($id);
 
         $validator = Validator::make($input, [
             'name' => 'required',
@@ -84,6 +89,53 @@ class ProductController extends BaseController
 
         $product->name = $input['name'];
         $product->detail = $input['detail'];
+
+        if (isset($input['code'])) {
+            $product->code = $input['code'];
+        }
+        if (isset($input['department_id'])) {
+            $product->department_id = $input['department_id'];
+        }
+        if (isset($input['category_id'])) {
+            $product->category_id = $input['category_id'];
+        }
+        if  (isset($input['image'])) {
+            $product->image = $input['image'];
+        }
+        if (isset($input['price'])) {
+            $product->price = $input['price'];
+        }
+        if (isset($input['status'])) {
+            $product->status = $input['status'];
+        }
+        if (isset($input['taxes'])) {
+            $product->taxes = $input['taxes'];
+        }
+        if (isset($input['maximum'])) {
+            $product->maximum = $input['maximum'];
+        }
+        if (isset($input['minimum'])) {
+            $product->minimum = $input['minimum'];
+        }
+        if (isset($input['unit'])) {
+            $product->unit = $input['unit'];
+        }
+        if (isset($input['profit'])) {
+            $product->profit = $input['profit'];
+        }
+        if (isset($input['sale_price'])) {
+            $product->sale_price = $input['sale_price'];
+        }
+        if (isset($input['measurement'])) {
+            $product->measurement = $input['measurement'];
+        }
+        if  (isset($input['description'])) {
+            $product->description = $input['description'];
+        }
+        if (isset($input['profit'])) {
+            $product->profit = $input['profit'];
+        }
+
         $product->save();
 
         return $this->sendResponse(new ProductResource($product), 'Product updated successfully.');
@@ -92,11 +144,18 @@ class ProductController extends BaseController
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product)
+    public function destroy(int $id)
     {
+
+        $product = Product::find($id);
+
+        if  (is_null($product)) {
+            return $this->sendError('Department not found.');
+        }
+
         $product->delete();
 
         return $this->sendResponse([], 'Product deleted successfully.');
